Issue #103: util.state: Re-structure for pre-3.3 Python.
Python older than v3.3 does not understand `yield from`, and won't
ignore it inspite of it being in a code path that won't be run for older
releases.

So instead, we have to put it in a completely separate file and convince
Python to hopefully not import it when running on older Python releases.
